stiff

+++ app/Http/Controllers/Api/SupplierController.php
@@ -75,6 +75,12 @@ class SupplierController extends Controller
         if (!$supplier) {
             return response()->json(['message' => 'Supplier not found'], 404);
         }
+        if($supplier->name=='system'){
+            return response()->json(['message'=>'the system supplier cannot be deleted']);
+        }
+            if($supplier->supplierOrders()->count()>0 || $supplier->supplierInvoices()->count()>0){
+                return response()->json(['message'=>'Cannot delete supplier with associated orders or invoices'],403);
+            }
         try{
         $supplier->delete();
         }catch(\Exception $e){
